fixed minibar filter problem!
see details - https://github.com/wdecoster/nanofilt/issues/40
the problem was the empty line reads in minibar outputs - added a first pass NanoFilt just on min len. the problem is that if you filter on quality and length at the same time, NanoFilt tries to calc avg quality from quality scores for every read. Therefore, if the read is 0 length, it errors out on the ave_qual( ) calculation.